pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cross-lingual-phoneme-aware-speech-enhancement-with-adaptive-masking"
version = "0.1.0"
description = "Multi-stage speech enhancement system that leverages cross-lingual phoneme embeddings to guide adaptive time-frequency masking for noise reduction in low-resource languages"
authors = [{name = "Alireza Shojaei"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyyaml>=6.0",
    "tensorboard>=2.12.0",
    "mlflow>=2.10.0",
    "pydub>=0.25.0",
    "pesq>=0.0.4",
    "pystoi>=0.3.3",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    "transformers>=4.30.0",
    "phonemizer>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
